// src/app/pages/items/rich-essay/model/rich-essay-model.model.ts
import { ItemTagsDtos } from '../../models/item-tags-dtos';
import { Images } from './../../models/images';
import { ScoringOption } from './../../models/scoring-option';

export class RichEssayModel {
  stimulus: string = '';
  itemId: string;
  reference: string = '';
  difficultyLevel: number = 1;
  scoringOption: ScoringOption = new ScoringOption();
  images: Images[] = [];
  passageId: string = null;
  itemTagsDTOS: ItemTagsDtos[] = [];
  subjectId: string;
  topicId: string;
  subtopicId: string;
  itemStatus: string;
  itemType: string;
  allowPaste: boolean;
  allowCopy: boolean;
  allowCut: boolean;
  plainText: boolean;
  maxWords: number;
  moderation_status: string = null
}